fix :) use context for UserId & 삭제 시에 본인 게시물만 삭제할 수 있도록 (관리자 제외)

+++ src/resolvers/mutation/studyBoard.js
@@ -1,14 +1,20 @@
 import models from "../../models";
 
 // 게시판 글 생성
-const createStudyBoard = async (_, { params }) => {
+const createStudyBoard = async (_, { params }, context) => {
+  params.UserId = context.user.id;
   const studyBoard = await models.StudyBoard.create(params);
 
   return studyBoard;
 };
 
 // 게시판 글 삭제
-const deleteStudyBoard = async (_, { id }) => {
+const deleteStudyBoard = async (_, { id }, context) => {
+  const user = context.user;
+  const studyBoard = await models.StudyBoard.findByPk(id);
+  if (!user.isAdmin && studyBoard.UserId !== user.id)
+    throw new Error("본인의 게시물만 삭제할 수 있습니다.");
+
   await models.StudyBoard.destroy({
     where: { id }
   });
